% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Export.R
\name{createAnalysisDataSet}
\alias{createAnalysisDataSet}
\title{Create analysis dataset}
\usage{
createAnalysisDataSet(dataFolders)
}
\arguments{
\item{dataFolders}{A vector of folder names where the cohort method analyses have
been executed.}
}
\value{
A tibble with one row per person, across all outcomes, analyses, and databases
}
\description{
Create analysis dataset
}
