% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  verifyDependencies = TRUE,
  outputFolder,
  databaseId = "Unknown",
  createCohorts = TRUE,
  runCohortMethod = TRUE,
  maxCores = 4
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write privilege in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{verifyDependencies}{Check whether correct package versions are installed?}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{createCohorts}{Create the cohortTable table with the exposure and outcome cohorts?}

\item{runCohortMethod}{Perform the cohort method analyses?}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Execute the Study
}
\details{
This function executes the Epi1055_1056 Study.
}
