% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionsEval.R
\docType{package}
\name{InteractionsEval-package}
\alias{InteractionsEval}
\alias{InteractionsEval-package}
\title{InteractionsEval: Evaluating Methods for Estimating Effect Interactions}
\description{
Evaluating methods for estimating effect interactions.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
